EmailAnsweingAIAgent/requirements.txt
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.125.0
google-generativeai==0.5.0
